Update function_tests.py
Note:
This file is within the Old folder, and changes could have been made to these functions before integration with other programs.  The more recent versions of these functions are in use within larger programs.  Use these separate test versions at your own risk.

Functions written and tested in this file include:
1. canonical_form
This function intakes a districting partition in a form produced by a separate enumeration function, and translates it into a standardized form, which the function returns.  This is important because we often need the computer to  be able to compare two districting plans and evaluate whether they are the same, but each plan has many possible relabelings.  Canonical form ensures that any two relabelings of the same districting plan are not mistaken for unique plans in future functions.  For more detail on canonical form, see the top-level ReadMe. 

For more detail on using pickles in python, see the top-level ReadMe.

2. get_enum_freqs
This function runs a sample of districting plans against the full enumeration of possible districting plans, with the goal of comparing the two sets.  For a fixed map, input a .csv of all possible enumerated districting plans under some constraints.  Input a sample of districting plans.  The function outputs a dictionary which stores how often each possible partition appeared in the sample.  
For example, 80 of the possible enumerated partitions could have appeared 0 times, while 5 of them appeared 4 times.   In order to evaluate the randomness of a sampling algorithm, we may be interested in understanding why certain partitions appear many times before others appear at all.  This function is only functional as long as full enumeration is feasible (not long at all).

The code at the end of this file plots the enum_freq output dictionary, to visualize whether there are any spikes in the graph for more commonly chosen partitionings.